# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Express Node.js API with MongoDB",
    "description": "This is a sample API for an Express Node.js application",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3002/api/v1",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/categories": {
      "get": {
        "summary": "Get a list of categories",
        "description": "Retrieve a list of categories with pagination support",
        "parameters": [
          {
            "in": "query",
            "name": "pageSize",
            "description": "Number of items per page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageNo",
            "description": "Page number",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of categories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new category",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewCategory"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          }
        }
      }
    },
    "/categories/{id}": {
      "get": {
        "summary": "Get a category by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "ID of the category to retrieve",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "404": {
            "description": "Category not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update a category by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "ID of the category to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewCategory"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": "Category could not be updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a category by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "ID of the category to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "404": {
            "description": "Category not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/orders": {
      "get": {
        "summary": "Get all orders",
        "responses": {
          "200": {
            "description": "List of orders",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "Create a new order",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewOrder"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Invalid order data"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "summary": "Get an order by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the order",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "summary": "Update an order status by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the order",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderStatus"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order status updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Invalid order data"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "summary": "Delete an order by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the order",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "The order is completely deleted."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid order ID"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/orders/get/totalSales": {
      "get": {
        "summary": "Get total sales amount",
        "responses": {
          "200": {
            "description": "Total sales amount",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "totalSales": {
                      "type": "number",
                      "example": 1000.5
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error getting total sales amount"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/orders/get/count": {
      "get": {
        "summary": "Get total order count",
        "responses": {
          "200": {
            "description": "Total order count",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "count": {
                      "type": "number",
                      "example": 10
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/orders/get/userOrder/{userId}": {
      "get": {
        "summary": "Get orders by user ID",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of orders by user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Category": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "image": {
            "type": "string"
          }
        }
      },
      "NewCategory": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "image": {
            "type": "string"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "Success": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "Order": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "607ebcb31703622ef09e6097"
          },
          "orderItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          },
          "shippingAddress1": {
            "type": "string",
            "example": "123 Main St"
          },
          "shippingAddress2": {
            "type": "string",
            "example": "Apt 101"
          },
          "city": {
            "type": "string",
            "example": "New York"
          },
          "country": {
            "type": "string",
            "example": "USA"
          },
          "zip": {
            "type": "string",
            "example": "10001"
          },
          "phone": {
            "type": "string",
            "example": "123-456-7890"
          },
          "status": {
            "type": "string",
            "example": "Pending"
          },
          "totalPrice": {
            "type": "number",
            "example": 99.99
          },
          "user": {
            "type": "string",
            "example": "607ebcb31703622ef09e6097"
          },
          "dateOrdered": {
            "type": "string",
            "format": "date-time",
            "example": "2022-04-21T11:23:10.760Z"
          }
        }
      },
      "NewOrder": {
        "type": "object",
        "properties": {
          "orderItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NewOrderItem"
            }
          },
          "shippingAddress1": {
            "type": "string",
            "example": "123 Main St"
          },
          "shippingAddress2": {
            "type": "string",
            "example": "Apt 101"
          },
          "city": {
            "type": "string",
            "example": "New York"
          },
          "country": {
            "type": "string",
            "example": "USA"
          },
          "zip": {
            "type": "string",
            "example": "10001"
          },
          "phone": {
            "type": "string",
            "example": "123-456-7890"
          },
          "status": {
            "type": "string",
            "example": "Pending"
          },
          "user": {
            "type": "string",
            "example": "607ebcb31703622ef09e6097"
          }
        }
      },
      "UpdateOrderStatus": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "Delivered"
          }
        }
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "607ebcb31703622ef09e6097"
          },
          "product": {
            "type": "string",
            "example": "607ebcb31703622ef09e6097"
          },
          "quantity": {
            "type": "number",
            "example": 2
          }
        }
      },
      "NewOrderItem": {
        "type": "object",
        "properties": {
          "product": {
            "type": "string",
            "example": "607ebcb31703622ef09e6097"
          },
          "quantity": {
            "type": "number",
            "example": 2
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}